Add initial version of subcommand framework
subcommand is a framework for creating subcommand-style programs (like
git, etc.) that allows you to use full power of bash.
Subcommands are simply functions in a script, and each function becomes
a subcommand.

There are still many TODOs and, probably, bugs, but it is working
somehow.

Added one example for now (examples/actl) that can be used as example
usage. Add it (symlink is OK) to your PATH, and run actl --help.